pyproject.toml
[project]
name = "spotify-archive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "python-dotenv>=1.0.1",
    "spotipy>=2.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]

[tool.ruff]
target-version = "py313"
line-length = 119
 
[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90"]
ignore = ["E203"]
